requirements.txt
cachetools==4.2.2
docopt==0.6.1
Flask==0.10.1
GitPython==0.3.2rc1
google-auth==2.0.0
google-auth-oauthlib==0.4.5
gspread==4.0.1
itsdangerous==2.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.7.2
semantic-version==2.3.0
Werkzeug==2.0.1
